backend/pyproject.toml
[project]
name = "karmafit-backend"
version = "0.1.0"
description = "Resume OCR backend using GOT-OCR"
requires-python = ">=3.10"

dependencies = [
#    "torch==2.3.1+cu121",
    "transformers @ git+https://github.com/huggingface/transformers.git",
    "accelerate>=0.30.0",
    "pillow>=10.0.0",
    "pdf2image>=1.17.0",
    "python-docx>=1.1.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    # Authentication & Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pydantic[email]>=2.0.0",
]

# [tool.uv]
# extra-index-url = ["https://download.pytorch.org/whl/cu121"]